requirements.txt
openai>=1.0.0
httpx>=0.24.0
rich>=13.0.0
prompt-toolkit>=3.0.0
python-dotenv>=1.0.0
tzdata>=2024.1  # Timezone data for Windows (zoneinfo needs this)
pyinstaller>=6.0.0
mcp>=0.1.0  # Model Context Protocol (optional, for tool support)
